+++ snippets.go
@@ -26,12 +26,12 @@ import (
 type (
 	SnippetsServiceInterface interface {
 		ListSnippets(opt *ListSnippetsOptions, options ...RequestOptionFunc) ([]*Snippet, *Response, error)
-		GetSnippet(snippet int, options ...RequestOptionFunc) (*Snippet, *Response, error)
-		SnippetContent(snippet int, options ...RequestOptionFunc) ([]byte, *Response, error)
-		SnippetFileContent(snippet int, ref, filename string, options ...RequestOptionFunc) ([]byte, *Response, error)
+		GetSnippet(snippet int64, options ...RequestOptionFunc) (*Snippet, *Response, error)
+		SnippetContent(snippet int64, options ...RequestOptionFunc) ([]byte, *Response, error)
+		SnippetFileContent(snippet int64, ref, filename string, options ...RequestOptionFunc) ([]byte, *Response, error)
 		CreateSnippet(opt *CreateSnippetOptions, options ...RequestOptionFunc) (*Snippet, *Response, error)
-		UpdateSnippet(snippet int, opt *UpdateSnippetOptions, options ...RequestOptionFunc) (*Snippet, *Response, error)
-		DeleteSnippet(snippet int, options ...RequestOptionFunc) (*Response, error)
+		UpdateSnippet(snippet int64, opt *UpdateSnippetOptions, options ...RequestOptionFunc) (*Snippet, *Response, error)
+		DeleteSnippet(snippet int64, options ...RequestOptionFunc) (*Response, error)
 		ExploreSnippets(opt *ExploreSnippetsOptions, options ...RequestOptionFunc) ([]*Snippet, *Response, error)
 		ListAllSnippets(opt *ListAllSnippetsOptions, options ...RequestOptionFunc) ([]*Snippet, *Response, error)
 	}
@@ -51,7 +51,7 @@ var _ SnippetsServiceInterface = (*SnippetsService)(nil)
 //
 // GitLab API docs: https://docs.gitlab.com/api/snippets/
 type Snippet struct {
-	ID                int           `json:"id"`
+	ID                int64         `json:"id"`
 	Title             string        `json:"title"`
 	FileName          string        `json:"file_name"`
 	Description       string        `json:"description"`
@@ -59,7 +59,7 @@ type Snippet struct {
 	Author            SnippetAuthor `json:"author"`
 	UpdatedAt         *time.Time    `json:"updated_at"`
 	CreatedAt         *time.Time    `json:"created_at"`
-	ProjectID         int           `json:"project_id"`
+	ProjectID         int64         `json:"project_id"`
 	WebURL            string        `json:"web_url"`
 	RawURL            string        `json:"raw_url"`
 	Files             []SnippetFile `json:"files"`
@@ -74,7 +74,7 @@ func (s Snippet) String() string {
 //
 // GitLab API docs: https://docs.gitlab.com/api/snippets/
 type SnippetAuthor struct {
-	ID        int        `json:"id"`
+	ID        int64      `json:"id"`
 	Username  string     `json:"username"`
 	Email     string     `json:"email"`
 	Name      string     `json:"name"`
@@ -129,7 +129,7 @@ func (s *SnippetsService) ListSnippets(opt *ListSnippetsOptions, options ...Requ
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/snippets/#get-a-single-snippet
-func (s *SnippetsService) GetSnippet(snippet int, options ...RequestOptionFunc) (*Snippet, *Response, error) {
+func (s *SnippetsService) GetSnippet(snippet int64, options ...RequestOptionFunc) (*Snippet, *Response, error) {
 	u := fmt.Sprintf("snippets/%d", snippet)
 
 	req, err := s.client.NewRequest(http.MethodGet, u, nil, options)
@@ -150,7 +150,7 @@ func (s *SnippetsService) GetSnippet(snippet int, options ...RequestOptionFunc) 
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/snippets/#single-snippet-contents
-func (s *SnippetsService) SnippetContent(snippet int, options ...RequestOptionFunc) ([]byte, *Response, error) {
+func (s *SnippetsService) SnippetContent(snippet int64, options ...RequestOptionFunc) ([]byte, *Response, error) {
 	u := fmt.Sprintf("snippets/%d/raw", snippet)
 
 	req, err := s.client.NewRequest(http.MethodGet, u, nil, options)
@@ -171,7 +171,7 @@ func (s *SnippetsService) SnippetContent(snippet int, options ...RequestOptionFu
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/snippets/#snippet-repository-file-content
-func (s *SnippetsService) SnippetFileContent(snippet int, ref, filename string, options ...RequestOptionFunc) ([]byte, *Response, error) {
+func (s *SnippetsService) SnippetFileContent(snippet int64, ref, filename string, options ...RequestOptionFunc) ([]byte, *Response, error) {
 	filepath := PathEscape(filename)
 	u := fmt.Sprintf("snippets/%d/files/%s/%s/raw", snippet, ref, filepath)
 
@@ -260,7 +260,7 @@ type UpdateSnippetOptions struct {
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/snippets/#update-snippet
-func (s *SnippetsService) UpdateSnippet(snippet int, opt *UpdateSnippetOptions, options ...RequestOptionFunc) (*Snippet, *Response, error) {
+func (s *SnippetsService) UpdateSnippet(snippet int64, opt *UpdateSnippetOptions, options ...RequestOptionFunc) (*Snippet, *Response, error) {
 	u := fmt.Sprintf("snippets/%d", snippet)
 
 	req, err := s.client.NewRequest(http.MethodPut, u, opt, options)
@@ -283,7 +283,7 @@ func (s *SnippetsService) UpdateSnippet(snippet int, opt *UpdateSnippetOptions, 
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/snippets/#delete-snippet
-func (s *SnippetsService) DeleteSnippet(snippet int, options ...RequestOptionFunc) (*Response, error) {
+func (s *SnippetsService) DeleteSnippet(snippet int64, options ...RequestOptionFunc) (*Response, error) {
 	u := fmt.Sprintf("snippets/%d", snippet)
 
 	req, err := s.client.NewRequest(http.MethodDelete, u, nil, options)
